+++ packages/examples/packages/cronjobs/src/index.test.ts
@@ -29,19 +29,4 @@ describe('onCronjob', () => {
       expect(response).toRespondWith(null);
     });
   });
-
-  describe('fireNotification', () => {
-    it('shows an inApp notification', async () => {
-      const { onBackgroundEvent } = await installSnap();
-
-      const response = await onBackgroundEvent({
-        // This would normally be called by the MetaMask extension, but to make
-        // this testable, `@metamask/snaps-jest` exposes a `onBackgroundEvent` method.
-        method: 'fireNotification',
-      });
-
-      expect(response).toRespondWith(null);
-      expect(response).toSendNotification('Hello world!', 'inApp');
-    });
-  });
 });