+++ packages/test-snaps/CHANGELOG.md
@@ -9,11 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Use background events example snap for testing background events.
+- Use background events example snap for testing background events ([#3404](https://github.com/MetaMask/snaps/pull/3404))
 
 ### Removed
 
-- Delete background events logic from cronjobs example snap section.
+- Delete background events logic from cronjobs example snap section ([#3404](https://github.com/MetaMask/snaps/pull/3404))
 
 ## [2.22.0]